Remove debug code and document bias issue
The simulation exhibits bias for vertex moves in the 111 and -1-1-1 directions when multiple threads access vertices concurrently via competitions for locks. This bias becomes more noticeable for many threads on an already-fast simulation. Ramp up number of threads for a given simulation to verify that this bias does not become an issue.